backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
pydantic-settings==2.2.1
supabase==2.11.0
PyJWT==2.8.0
cryptography==42.0.5
python-dotenv==1.0.1
python-multipart==0.0.9
slowapi==0.1.4
structlog==24.1.0
